Fix V logging to also filter out secrets
We use a small logging wrapper around glog for most of our logging. The
global Errorf/Warningf/Infof methods in that module format the string
and args passed in then filter out any secrets before calling into glog
with the final string.

The wrapper also exposed a `V` method to only log if verbosity was set,
but that directly returned a glog.Verbose which mean the Infof calls to
that didn't run our secret filtering code.

This changes the logging wrapper to return a `VerboseShim` (Not a great
name, but `Verbose` was already taken in this module) which is simply a
new type over `glog.Verbose` with the same methods but it calls
`FilterString` on the inputs.

Technically a breaking change, but I think the uses of this will be
source compatible.