sdk/go: Drop unused resourceOptions field
The `Provider` field is never set by any option.
The corresponding option `func Provider(..)`,
when used as a `ResourceOption` delegates to the `Providers` function
which sets the `Providers` map instead.

```go
func Provider(r ProviderResource) ResourceOrInvokeOption {
	... {
	    Providers(r).applyResourceOption(ro)
	}
}
```

Drop the unused field in preparation for exposing
a mirror of the resourceOptions struct for #11698.

Refs #11698