ResourceOptions: Document fields for external use
The documentation for the fields of resourceOptions
was largely copied 1-1 from the corresponding option's documentation.

This was fine while these fields were private,
but now that we're exporting ResourceOptions,
these fields should be documented from the point of view
of a consumer of the struct not the producer of corresponding options.