{cli, filestate}: Add 'state upgrade' command
Adds a new `pulumi state upgrade` command
that will upgrade the state for the current state storage
to the latest version supported by that format.
This is a no-op for httpstate,
but for filestate, this will migrate from unversioned storage
to the project-based versioned format.

The upgrade command is backed by a new Upgrade operation
on the filestate Backend.
The upgrade operation operates by effectively renaming stacks
from legacy store to the new project store.

UX notes:

- We'll warn about stacks we failed to upgrade
  but keep going otherwise.
- We'll print a loud warning before upgrading,
  informing users that the state will not be usable
  from older CLIs.

Extracted from #12134

Co-authored-by: Abhinav Gupta <abhinav@pulumi.com>