Try to delete temporary files from plugin downloads
Fixes https://github.com/pulumi/pulumi/issues/12144

There's only three places in the codebase that we call
`workspace.DownloadToFile`. Before this change only one of them tried to
run `os.Remove` to cleanup afterwards.

This unifies all to use `os.Remove` and to also explictly ignore the
error returned by that (if it does fail it will log to `V(3)`).