cli/pulumi import: Don't panic on invalid file
`pulumi import -f` panics if the JSON file provided as input
is missing information like ID or name.

This adds validation for the required fields of deploy.Import
when we're building it from the parsed file.

To provide maximum context in the error messages,
it attempts to extract any information it can glean
about the provided importSpec.

Example run:

```
% cat tmp.json
{
        "resources": [
                {
                        "type": "someTokenType",
                        "name": "someName"
                }
        ]
}

% pulumi import -f tmp.json
error: resource 'someName' of type 'someTokenType' has no ID
```

Resolves #12103