sdk/go: Fix panic from uninitialized parent resources
The Go SDK panics if RegisterResource is called with a parent resource
that is uninitialized
(hasn't had its RegisterResource or RegisterComponentResource call yet).

The panic is during alias collapsing:
the system fails to read the unintialized parent resource's URN
from a zero-valued URNOutput.

Guard against this panic by verifying that the value passed to `Parent`
has been initialized.
We do this by checking the name of the resource--names are not allowed
to be empty when a resource is registered.

Resolves #12138