Remove appendix from LICENSE
Fixes https://github.com/pulumi/pulumi/issues/12316

As per https://opensource.stackexchange.com/questions/4149/include-appendix-of-apache-2-0-in-license-file
the appendix of the Apache 2.0 license does not need to be included in
the LICENSE file. We keep getting people commenting that we haven't
filled in the name and data in the appendix so to reduce confusion lets
just remove it.