[cli] Fetch update tokens in the API client
Rather than fetching update tokens in consumers of the API client, fetch
them in the API client itself. This minimizes the latency between the
retrieval of the token and the validation of the token by the Pulumi
Service. This is intended to help prevent the use of expired tokens.

Part of #7094.