ci: Run action-install-gh-release with auth
We have a few places in CI where we use action-install-gh-release
to download and install specific or latest binary releases of a
dependency.

Nearly all these steps, besides the one in ci-test-codegen,
were unauthenticated.
This caused us to run into rate limits with the GitHub API
when trying to get the latest release and resulted in delays like:

```
Run jaxxstorm/action-install-gh-release@v1.7.1
==> System reported platform: linux
==> Using platform: linux
==> System reported arch: x64
==> Using arch: x64
Warning: RateLimit detected for request GET /repos/{owner}/{repo}/releases/tags/{tag}.
Retrying after 3300 seconds.
Error: The operation was canceled.
```

This sets GITHUB_TOKEN for all instances of this step
so that we're less likely to get rate limited by GitHub on this.