Add hidden --overlays option to gen-sdk
This is hidden because we don't really want people using overlays,
the current plan is try and remove them all. But we currently have some
in use in existing providers and we need to support that to migrate
those providers over to using gen-sdk rather than calling into the
codegen libraries themselves.