ci: Don't run codegen tests for PRs
We currently run all codegen tests in pkg/codegen/$lang
for every PR.
These tests take quite a while to run and lock up many GitHub workers
for this entire duration.

This change attempts to address this issue
by running codegen tests only for those PRs
that touch the codegen directories.

The machinery to make this work is roughly as follows:

- In the on-pr workflow, when we're figuring out what we're doing
  we label the PR with ci/test-codegen.
  We use [actions/labeler] to do this.
  The name of the label matches "ci/test" which tells the system
  to run all integration and unit tests.
- Using the label, we decide whether we want to run codegen tests,
  and pass that onto the test matrix generator.
- The test matrix generator filters out these packages
  and their subpackages from the list of tests under consideration.
- Everything else proceeds as normal.

Things to note:

- The codegen-tests input defaults to true.
  All other invocations will run with code gen tests
  so these will continue to run on merge.
  Only PRs (from on-pr.yml) set it to false.
- Since the number of tests is remarkably smaller without these tests,
  we can significantly reduce the number of partitions we use
  for pkg/ unit tests.
  This should alleviate pressure on GitHub workers further.

The choice to use [actions/labeler] was made because GitHub
does not support filtering on files in the pull request
inside `if` clauses or other `${{ .. }}` expressions.

  [actions/labeler]: https://github.com/actions/labeler

This is a pretty blunt approach to the problem.
If we wanted to be more targeted,
instead of filtering at the get-job-matrix.py level,
we could instead set an environment variable
and add t.Skips in {program,sdk,type}_driver
if that environment variable is set.

Resolves #12334