refactor(cmd/whoami): Turn into a command struct
Moves the core logic of the whoami command into a whoAmICmd struct,
with the flags as fields on the struct.

This is a step towards making this code tesetable echoing the change
made to `stack export` in fa7bccbc903298e0949b46c603cb81c992f650de.

This commit no logic changes.