Resources now inherit the `DeletedWith` Resource Option across SDKs.
This is implemented in the engine and interprets the empty string `""`
to inherit the value from the resource's parent if it exists signifying
that it was unspecified by the user's program. There is currently no way
to override this in a child to unset it when set by the parent, but can
be addressed by not parenting the resource to a resource with
`deletedWith` set.