localBackendReference: Don't track backend
localBackendReference keeps track of the backend
so that it can determine the current project
and shorten the String representation of the reference.

As far as I can tell,
this doesn't need to be "current project at the time of calling String"
and can be "current project when the reference was constructed."
This switch allows us to turn localBackendReference
back into a data-only object.