referenceStore: More control over layout, track on localBackendReference
Let referenceStore decide paths of different files and directories
rather than having the backend decide them
based on whether localBackendReference.project is empty or not.

To make renames from legacy to project-based stacks possible,
we need localBackendReference to track the store that created it,
and use that to determine old and new paths.