test(filestate): Add legacy copies of migrated tests
The filestate backend now defaults to using project mode for new stores.
This has a side effect of migrating stores in all existing tests
to project mode.
As a result of this, we lost some of the legacy test.

To get around this, create legacy copies of affected tests.
We run them in legacy mode by ensuring that the bucket is non-empty
with a marker file.