filestate/projectStore: Optimize ListReferences
ListReferences for project layout is currently inefficient.
It roughly does the equivalent of:

    for proj in bucket.List(stacksDir)
        for ref in bucket.List(stacksDir + "/" + proj):
            # ...

This is fine for local, file-based storage,
but super inefficient for S3 or Azure.
The reason this was written as above is because we treat the bucket
as a directory tree in listBucket.

This changes projectStore.ListReferences to use the underlying
bucket.List method directly--without setting the Delimiter.
This treats the bucket as a flat file list instead of a hierarchy,
and we can extract project and stack names from it in one go.