refactor(sdk/workspace/DownloadToFile): Pull state into a struct
All state for DownloadToFile was in a bunch of anonymous functions
that called each other.
The only way to inject new options was via new positional parameters.

This turns the core logic of DownloadToFile into
a pluginDownloader struct,
with retry and wrapper injected as fields rather than parameters.
It further breaks out all anonymous functions in DownloadToFile
into methods on that struct.
Their order has been retained to make this easy to review.

This change contains no logic changes.