sdk/workspace/pluginDownloader: Re-use util/retry
The retry logic wants to make at most 5 attempts
(that's what the "limit" says)
but it actually makes 6 requests.

This fixes the off-by-one error
by switching to the re-usable util/retry package
instead of re-implementing retries here.

As with util/retry.Retryer,
pluginDownloader allows overriding time.After from tests,
so that we don't sleep in tests.