sdk/common/retry: Allow injecting time.After, add tests
This extracts the core logic in the retry package into a Retryer struct.
The Retryer struct supports an optional "After" field,
which allows injecting a custom time.After implementation.

With this, we're able to write tests for this package
without actually sleeping (and slowing down the test suite).

Note that the tests exposed two surprising behaviors:

- If the context times out or expires, we correctly return `false`,
  but we return a nil error instead of `ctx.Err()`.
  I've chosen to not change this at this time
  because we don't know what that might break downstream.
- The initial delay is never used. We start with delay*backoff.