lint/revive: Ignore redefines-builtin-id for 'new'
redefines-builtin-id is useful enough as a stylistic aid.
The one case where the value of the linter is questionable
is with 'new'.

We use 'new' frequently as a variable name,
especially in the context of diffs (old and new).
We rarely use the `new` built-in function.
So we opt out of this check only for 'new'.