cmd[cancel, convert, env, poly rm, stack rm]: Use ErrBail
These command implementations are simple enough
to easily switch to RunResult and ErrBail
instead of needlessly using result.Result.

For a couple of the commands,
this also lets us put the multi-line, formatted error message
in an error log entry with cmdutil.Diag().Errorf,
and bail cleanly.
This will give us a nice color-coded message for those cases.