[sdk/python] Fix enum test for python 11 behavior change
Python 11 has changed the behavior of `__format__` and `__str__` to return the type and name of an enum rather than its value. This affects one of our tests which is expecting the value of the enum ("Burgundy") to be formatted in a sentence and not the type and name ("RubberTreeVariety.BURGUNDY"). This change fixes the test to explicitly use the enum's value in the formatted string so we get consistent output when the test is run across multiple versions of Python.