filestate: Warn about mixed states
It's possible to end up in a mixed state
where some stacks are project-scoped and others aren't.

For instance, imagine the scenario:

- *User A* sets up a new state with the latest version of the CLI
  and brings up a new stack
- *User B* brings up a stack in that state using an older version;
  the older version doesn't know about the .pulumi/meta.yaml
  so it doesn't explode and writes a legacy stack file

At this point, the state is in mixed state.

To help users detect (and eventually resolve) this issue,
we'll print a warning when the backend is initialized.
It'll find and warn users about any legacy files.
They can opt out of this check by setting an environment variable.

This will help transition users to project-scoped stacks
until we're ready to delete this environment variable
and the check.