[macOS] cli: Suggest brew update in upgrade message
On macOS, the message informing users
that a new version of Pulumi is available
takes the form:

```
warning: A new version of Pulumi is available. To upgrade from version '3.58.0' to '3.59.0', run
   $ brew upgrade pulumi
```

If the user has not run `brew update` in a while,
this will not actually perform the upgrade.

Make the command more copy-paste friendly
by also suggesting a `brew update`.