Temporary compatability shim for tf mappings
Currently bridge provider plugins return a mapping for "tf" but not
"terraform", the bridge has been updated to reply for both but given it
takes time for providers to be updated and published we're adding this
compat shim to just try both names.

Also closes off the provider plugins after calling GetMapping to stop a
resource leak.