refactor(cmd/stack init): Turn into a command struct
This refactors the `pulumi stack init` command implementation
in the same vein as #12400 and #11951.

In short, this moves
flags and global dependencies are moved to a stackInitCmd struct,
and the command implementation into a Run method on that struct.

Doing so allows direct testing of the command implementation.
As demonstration, this includes a variant of the
"--teams is not supported" test directly against the command.

This contains no behavior changes.
However, this does move the default value of the --secrets-provider flag
into the Run method instead of being applied by Cobra.