sdk/workspace/pluginDownloader: Fix multiplied retries
The pluginDownloader retries downloads with specialized logic,
and it relies on getHTTPResponse to make the HTTP requests.
However, getHTTPResponse retries the requests *it* makes as well.

This changes getHTTPResponse to not perform any retries,
and makes a copy of it, getHTTPResponseWithRetry, that does.
The pluginDownloader uses getHTTPResponse,
and other prior uses of getHTTPResponse have switched to the new one.

This resolves the multiplied retries for pluginDownloader,
however there's still an off-by-one error.