filestate: Don't write metadata for legacy layouts
In #12472, we added a new file to filestate backends:
.pulumi/Pulumi.yaml.
This file is intended to store metadata about the layout
in anticipation of support for project-scoped stacks.

This had the unintended side-effect of breaking users
who use tight access control on their S3 buckets:
users don't get write access outside specific stack directories
so they get an error like the following:

    error: write ".pulumi/Pulumi.yaml": blob (key ".pulumi/Pulumi.yaml") (code=Unknown): AccessDenied: Access Denied

This changes filestate.New to never write the file
if the version is zero
so users that don't currently have write access to that directory
will continue to be able to use the backend.

Resolves #12534