filestate: Rename Pulumi.yaml to meta.yaml
In #12472, we added a .pulumi/Pulumi.yaml file to filestate backends
to hold metadata about the storage layout format.
It has been brought up multiple times that this is a confusing name,
so we're renaming it to meta.yaml.

Backwards compatibility:
Although we've already shipped #12472,
and have started writing Pulumi.yaml files,
we don't currently use it for anything.
Its first use will be in #12437 when we add project-scoped stacks.
Therefore, this change is completely backwards compatible.

We'll end up leaving the extraneous .pulumi/Pulumi.yaml file behind
if the bucket was initialized with Pulumi v3.59.0 or newer,
before this change was released. This is not a high cost.
I don't think we should go back and delete these files;
I'm reluctant for the backend to perform destructive operations
for files that are technically not managed by it.