filestate/Upgrade: Upgrade meta.yaml before the rest
We currently upgrade the meta.yaml *after* upgrading all stack files.
This has a risk of leaving a state store inaccessible.
For example, if someone only has write access to `.pulumi/*/dev`,
and not `.pulumi/*` (#12534), the system would move the stack files,
and fail to write a meta.yaml.
This would leave the stored information completely inaccessible
because lack of meta.yaml would indicate legacy layout,
but the files have all been migrated to the project layout.

This changes filestate.Upgrade to upgrade the meta.yaml first.
If the operation fails, we don't try anything else.