Check if Plugin is bundled before installing
This commit adds a check to `pulumi plugin install` to confirm
that the plugin is not bundled with Pulumi before performing the download.
Without this check, the CLI will produce a 404 error because the language
plugins aren't distributable independently of the CLI executable.

If PULUMI_DEV is set we'll skip the error and try to download anyway.