filestate/lock: Fix path matching on Windows
Integration tests for filestate fail on Windows
because the stack reference path is a mix of `/` and `\\` symbols.
e.g.

```
want: .pulumi/locks/organization/testing-config/testing/640b9586-d794-4cce-a586-9ab58549d3d3.json
 got: .pulumi/locks/organization\testing-config\testing/640b9586-d794-4cce-a586-9ab58549d3d3.json
```

On Windows, the `<org>/<proj>/<name>` combo
gets turned to `<org>\<proj>\<name>` by `stackLockDir(stackName)`.
This is normally not a problem
because the result is fed into the blob.Bucket, which handles both,
but here we do exact string matching where this distinction matters.

To fix, make sure to convert the path to `/` form before comparison.